requirements.txt
python-dotenv==0.20.0
requests==2.26.0
Flask==2.1.2
locust==2.8.6
pytest==7.1.2